--- src/com/AmigoInvisible/isst/dao/EMFService.java
@@ -1,17 +1,0 @@
-package com.AmigoInvisible.isst.dao;
-
-import javax.persistence.EntityManagerFactory;
-import javax.persistence.Persistence;
-
-public class EMFService {
-	private static final EntityManagerFactory emfInstance = Persistence.createEntityManagerFactory("transactions-optional");
-	
-	private EMFService(){
-	}
-	
-	public static EntityManagerFactory get(){
-		return emfInstance;
-	}
-	
-
-}
