+++ src/es/upm/dit/isst/amigos/dao/AgrupacionesDAO.java
@@ -1,0 +1,17 @@
+package es.upm.dit.isst.amigos.dao;
+
+import java.util.List;
+
+import es.upm.dit.isst.amigos.model.Agrupaciones;
+
+public interface AgrupacionesDAO {
+	
+	// Devuelve el model, por si es necesario algun dato autogenerado.
+	public Agrupaciones insertAgrupacion(String user, int grupo, String amigoinv);
+	
+	public List<Agrupaciones> getAgrupacionesByUser(String user);
+	
+	public List<Agrupaciones> getAgrupacionesByGrupo(int grupo);
+	
+	public List<Agrupaciones> getAgrupByUserAndGrupo(String user, int grupo);
+}
