+++ src/es/upm/dit/isst/amigos/dao/EMFService.java
@@ -1,0 +1,17 @@
+package es.upm.dit.isst.amigos.dao;
+
+import javax.persistence.EntityManagerFactory;
+import javax.persistence.Persistence;
+
+public class EMFService {
+	private static final EntityManagerFactory emfInstance = Persistence.createEntityManagerFactory("transactions-optional");
+	
+	private EMFService(){
+	}
+	
+	public static EntityManagerFactory get(){
+		return emfInstance;
+	}
+	
+
+}
